Refactor UpdateProduct to use route id, use decimal for price
- Update UpdateProduct endpoint to accept product id from route instead of DTO
- Remove Id property from UpdateProductDto
- Adjust ProductMapper to map id from route to command
- Change Price type from int to decimal in DTOs and commands for accuracy

+++ Services/Catalog/Catalog.Application/DTOs/CreateProductDto.cs
@@ -24,5 +24,5 @@ public record class CreateProductDto
 
     [Required]
     [Range(0.01, double.MaxValue, ErrorMessage = "Price must be greater than zero.")]
-    public int Price { get; init; }
+    public decimal Price { get; init; }
 }
